moves common destination module to separate folder
