+++ dlt/common/destination/__init__.py
@@ -1,0 +1,1 @@
+from dlt.common.destination.capabilities import DestinationCapabilitiesContext, TLoaderFileFormat
