+++ dlt/common/destination/reference.py
@@ -1,57 +1,17 @@
 from abc import ABC, abstractmethod
 from importlib import import_module
 from types import TracebackType, ModuleType
-from typing import Any, Callable, ClassVar, Final, List, Optional, Literal, Type, Protocol, Union, TYPE_CHECKING, cast
+from typing import ClassVar, Final, Optional, Literal, Type, Protocol, Union, TYPE_CHECKING, cast
 
 from dlt.common.configuration.utils import serialize_value
 from dlt.common.exceptions import IdentifierTooLongException, InvalidDestinationReference, UnknownDestinationModule
 from dlt.common.schema import Schema
 from dlt.common.schema.typing import TTableSchema
+from dlt.common.schema.exceptions import InvalidDatasetName
 from dlt.common.configuration import configspec
-from dlt.common.configuration.specs import BaseConfiguration, CredentialsConfiguration, ContainerInjectableContext
+from dlt.common.configuration.specs import BaseConfiguration, CredentialsConfiguration
 from dlt.common.configuration.accessors import config
-
-
-# known loader file formats
-# jsonl - new line separated json documents
-# puae-jsonl - internal extract -> normalize format bases on jsonl
-# insert_values - insert SQL statements
-TLoaderFileFormat = Literal["jsonl", "puae-jsonl", "insert_values"]
-
-
-@configspec(init=True)
-class DestinationCapabilitiesContext(ContainerInjectableContext):
-    """Injectable destination capabilities required for many Pipeline stages ie. normalize"""
-    preferred_loader_file_format: TLoaderFileFormat
-    supported_loader_file_formats: List[TLoaderFileFormat]
-    escape_identifier: Callable[[str], str]
-    escape_literal: Callable[[Any], Any]
-    max_identifier_length: int
-    max_column_identifier_length: int
-    max_query_length: int
-    is_max_query_length_in_bytes: bool
-    max_text_data_type_length: int
-    is_max_text_data_type_length_in_bytes: bool
-    supports_ddl_transactions: bool
-
-    # do not allow to create default value, destination caps must be always explicitly inserted into container
-    can_create_default: ClassVar[bool] = False
-
-
-def generic_destination_capabilities(preferred_loader_file_format: TLoaderFileFormat = None) ->DestinationCapabilitiesContext:
-    caps = DestinationCapabilitiesContext()
-    caps.preferred_loader_file_format = preferred_loader_file_format
-    caps.supported_loader_file_formats = ["jsonl", "insert_values"]
-    caps.escape_identifier = lambda x: x
-    caps.escape_literal = lambda x: serialize_value(x)
-    caps.max_identifier_length = 65536
-    caps.max_column_identifier_length = 65536
-    caps.max_query_length = 32 * 1024 * 1024
-    caps.is_max_query_length_in_bytes = True
-    caps.max_text_data_type_length = 1024 * 1024 * 1024
-    caps.is_max_text_data_type_length_in_bytes = True
-    caps.supports_ddl_transactions = True
-    return caps
+from dlt.common.destination.capabilities import DestinationCapabilitiesContext
 
 
 @configspec(init=True)
@@ -170,6 +130,22 @@ class JobClientBase(ABC):
                         f"{table_name}.{column_name}",
                         self.capabilities.max_column_identifier_length
                     )
+
+    @staticmethod
+    def make_dataset_name(schema: Schema, dataset_name: str, default_schema_name: str) -> str:
+        """Builds full db dataset (dataset) name out of (normalized) default dataset and schema name"""
+        if not schema.name:
+            raise ValueError("schema_name is None or empty")
+        if not dataset_name:
+            raise ValueError("dataset_name is None or empty")
+        norm_name = schema.naming.normalize_identifier(dataset_name)
+        if norm_name != dataset_name:
+            raise InvalidDatasetName(dataset_name, norm_name)
+        # if default schema is None then suffix is not added
+        if default_schema_name is not None and schema.name != default_schema_name:
+            norm_name += "_" + schema.name
+
+        return norm_name
 
 
 TDestinationReferenceArg = Union["DestinationReference", ModuleType, None, str]
